Final Adjustment
Deleting unused activities & fragment
Change text from English to Bahasa Indonesia
