+++ app/src/main/AndroidManifest.xml
@@ -1,37 +1,32 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.example.capstone">
-    <uses-feature android:name="android.hardware.camera.any" />
-    <uses-permission android:name="android.permission.CAMERA" />
-    <uses-permission android:name="android.permission.INTERNET" />
-    <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE"/>
-    <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>
+
     <application
         android:allowBackup="true"
-        android:icon="@mipmap/ic_launcher"
+        android:icon="@drawable/leafwell"
         android:label="@string/app_name"
-        android:roundIcon="@mipmap/ic_launcher_round"
+        android:roundIcon="@drawable/leafwell"
         android:supportsRtl="true"
         android:theme="@style/Theme.Capstone">
-        <activity android:name=".ui.gallery.GalleryActivity"/>
-        <activity android:name=".ui.register.RegisterActivity" />
-        <activity android:name=".ui.camera.CameraActivity" />
-        <activity android:name=".ui.login.LoginActivity"/>
-        <activity android:name=".ui.detail.DetailActivity" />
-        <activity android:name=".ui.searchmanual.SearchManualActivity" />
-        <activity android:name=".MlActivity"/>
         <activity
-            android:name=".MainActivity"
-            android:label="@string/app_name"
-            android:theme="@style/Theme.Capstone.NoActionBar">
+            android:name=".ui.home.HomeActivity"
+            android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".ui.detail.DetailActivity" />
+        <activity android:name=".ui.searchmanual.SearchManualActivity" />
+        <activity android:name=".MlActivity" />
+        <activity android:name=".MainActivity">
+
+        </activity>
     </application>
 
-
+    <uses-permission android:name="android.permission.INTERNET" />
+    <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
 
 </manifest>
