--- app/src/main/java/com/example/capstone/data/LoginService.kt
@@ -1,19 +1,0 @@
-package com.example.capstone.data
-
-
-import retrofit2.Call
-import retrofit2.http.*
-
-
-
-interface LoginService {
-
-    @FormUrlEncoded
-    @POST("users/")
-    fun userLogin(
-      @Field("email") email : String,
-      @Field("username") username : String
-    ): Call<UserResponse>
-
-
-}
