--- app/src/main/java/com/example/capstone/data/SignUpResponse.kt
@@ -1,5 +1,0 @@
-package com.example.capstone.data
-
-data class SignUpResponse(
-    val user: User,
-)
