--- app/src/main/java/com/example/capstone/data/UserRequest.kt
@@ -1,16 +1,0 @@
-package com.example.capstone.data
-
-import com.google.gson.annotations.Expose
-import com.google.gson.annotations.SerializedName
-
-class UserRequest {
-
-    @SerializedName("email")
-    @Expose
-    var email: String?= null
-
-    @SerializedName("password")
-    @Expose
-    var password: String? =null
-
-}
