--- app/src/main/java/com/example/capstone/data/UserResponse.kt
@@ -1,7 +1,0 @@
-package com.example.capstone.data
-
-data class UserResponse (
-
-  val user : User,
-) 
-
