+++ app/src/main/java/com/example/capstone/ui/detail/DetailActivity.kt
@@ -1,9 +1,8 @@
 package com.example.capstone.ui.detail
 
-import android.media.Image
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.widget.Toast
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.view.isVisible
 import androidx.lifecycle.ViewModelProvider
 import com.bumptech.glide.Glide
 import com.bumptech.glide.request.RequestOptions
@@ -20,7 +19,6 @@ class DetailActivity : AppCompatActivity() {
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
-        var isFavorite = false
         binding = ActivityDetailBinding.inflate(layoutInflater)
         super.onCreate(savedInstanceState)
         setContentView(binding.root)
@@ -35,23 +33,14 @@ class DetailActivity : AppCompatActivity() {
             if (herbId != null) {
                 viewModel.getHerb(herbId)
                 viewModel.getHerbDetail().observe(this, { detail ->
+                    binding.progressBar.isVisible = false
                     populateHerb(detail)
-
                 })
 
             }
         }
         supportActionBar?.title = "Detail"
-//        binding.ivFavorite.setOnClickListener {
-//            isFavorite = !isFavorite
-//            if (isFavorite) {
-//                binding.ivFavorite.setBackgroundResource(R.drawable.ic_favorite_true_24)
-//                Toast.makeText(this, "Favorited", Toast.LENGTH_SHORT).show()
-//            } else {
-//                binding.ivFavorite.setBackgroundResource(R.drawable.ic_favorite_false_24)
-//                Toast.makeText(this, "Not Favorite", Toast.LENGTH_SHORT).show()
-//            }
-//        }
+
     }
 
     private fun populateHerb(detail: DetailEntity) {
