--- app/src/main/java/com/example/capstone/ui/register/RegisterActivity.kt
@@ -1,92 +1,0 @@
-package com.example.capstone.ui.register
-
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.widget.Toast
-import com.example.capstone.MainActivity
-import com.example.capstone.data.network.ApiConfig
-import com.example.capstone.data.SharedPrefManager
-import com.example.capstone.data.SignUpResponse
-import com.example.capstone.databinding.ActivityRegisterBinding
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
-
-class RegisterActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityRegisterBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        binding = ActivityRegisterBinding.inflate(layoutInflater)
-        super.onCreate(savedInstanceState)
-        setContentView(binding.root)
-
-        binding.btnRegister.setOnClickListener {
-
-            val username = binding.etEmail.toString().trim()
-            val password = binding.etPassword.toString().trim()
-            val date = binding.etUsername.toString().trim()
-
-            if (username.isEmpty()) {
-                binding.etEmail.error = "Username required"
-                binding.etEmail.requestFocus()
-                return@setOnClickListener
-            }
-
-
-            if (password.isEmpty()) {
-                binding.etPassword.error = "Password required"
-                binding.etPassword.requestFocus()
-                return@setOnClickListener
-            }
-
-            if (date.isEmpty()) {
-                binding.etUsername.error = "Date required"
-                binding.etUsername.requestFocus()
-                return@setOnClickListener
-            }
-            ApiConfig.createUser().createUser(username, date, password)
-                .enqueue(object : Callback<SignUpResponse> {
-                    override fun onResponse(
-                        call: Call<SignUpResponse>,
-                        response: Response<SignUpResponse>
-                    ) {
-                        if (response.isSuccessful) {
-                            Toast.makeText(
-                                applicationContext,
-                                "Sign Up berhasil",
-                                Toast.LENGTH_LONG
-                            ).show()
-                            startActivity(Intent(this@RegisterActivity, MainActivity::class.java))
-                            finish()
-                        } else {
-                            Toast.makeText(
-                                applicationContext,
-                                "Maaf ulangi kembali",
-                                Toast.LENGTH_LONG
-                            ).show()
-                        }
-
-                    }
-
-                    override fun onFailure(call: Call<SignUpResponse>, t: Throwable) {
-                        Toast.makeText(applicationContext, t.message, Toast.LENGTH_LONG).show()
-                    }
-
-                })
-
-
-        }
-    }
-
-    override fun onStart() {
-        super.onStart()
-
-        if(SharedPrefManager.getInstance(this).isLoggedIn){
-            val intent = Intent(applicationContext, MainActivity::class.java)
-            intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
-
-            startActivity(intent)
-        }
-    }
-}
