+++ app/src/main/res/layout/activity_detail.xml
@@ -13,6 +13,17 @@
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintStart_toStartOf="parent">
 
+        <ProgressBar
+            android:id="@+id/progress_bar"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:foregroundGravity="center"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            android:visibility="visible"/>
+
         <FrameLayout
             android:id="@+id/fl_backdrop"
             android:layout_width="match_parent"
@@ -26,17 +37,10 @@
                 android:layout_height="200dp"
                 app:layout_constraintTop_toTopOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
-                android:background="@color/blue"
+                android:background="@color/gray_white"
                 android:layout_gravity="center_horizontal"
                 android:scaleType="centerCrop"/>
 
-<!--            <ImageView-->
-<!--                android:id="@+id/iv_favorite"-->
-<!--                android:layout_width="30dp"-->
-<!--                android:layout_height="30dp"-->
-<!--                android:background="@drawable/ic_favorite_false_24"-->
-<!--                android:layout_gravity="bottom|right"-->
-<!--                android:layout_margin="20dp"/>-->
         </FrameLayout>
 
         <LinearLayout
@@ -55,7 +59,6 @@
                 android:layout_height="wrap_content"
                 app:layout_constraintTop_toBottomOf="@id/fl_backdrop"
                 app:layout_constraintStart_toStartOf="parent"
-                android:text="TitleGoesHere"
                 android:textStyle="bold"
                 android:textSize="24sp"
                 android:layout_marginBottom="18dp"/>
@@ -65,7 +68,7 @@
                 android:layout_height="wrap_content"
                 app:layout_constraintTop_toBottomOf="@id/fl_backdrop"
                 app:layout_constraintStart_toStartOf="parent"
-                android:text="Description"
+                android:text="@string/deskripsi"
                 android:textStyle="bold"
                 android:textSize="18sp"/>
 
@@ -81,7 +84,7 @@
                 android:layout_height="wrap_content"
                 app:layout_constraintTop_toBottomOf="@id/fl_backdrop"
                 app:layout_constraintStart_toStartOf="parent"
-                android:text="Efficacy"
+                android:text="@string/khasiat"
                 android:textStyle="bold"
                 android:textSize="18sp"/>
 
@@ -96,8 +99,8 @@
                 android:layout_height="wrap_content"
                 app:layout_constraintTop_toBottomOf="@id/fl_backdrop"
                 app:layout_constraintStart_toStartOf="parent"
-                android:text="How to Make It"
                 android:textStyle="bold"
+                android:text="@string/cara_mengolah"
                 android:textSize="18sp"/>
 
             <TextView
