+++ app/src/main/res/layout/activity_ml.xml
@@ -1,10 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".ui.home.HomeActivity">
 
     <TextView
         android:layout_width="wrap_content"
@@ -13,9 +12,11 @@
         android:layout_above="@+id/image"
         android:layout_centerHorizontal="true"
         android:layout_marginBottom="30dp"
-        android:text="Click below to select a Image"
+        android:text="@string/click_below_to_select_a_image"
+        android:paddingHorizontal="50dp"
+        android:gravity="center"
         android:textStyle="bold"
-        android:textSize="20dp">
+        android:textSize="24sp">
 
     </TextView>
 
@@ -39,7 +40,7 @@
         android:layout_marginEnd="50dp"
         android:layout_marginBottom="0dp"
         android:background="#fff"
-        android:text="Classify"
+        android:text="@string/klasifikasi"
         android:textColor="@color/white"
         android:enabled="false"
         android:textSize="18sp"
