--- app/src/main/res/layout/fragment_choice.xml
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.home.choicedialog.ChoiceFragment"
-    android:orientation="horizontal"
-    android:padding="20dp">
-
-    <TextView
-        android:id="@+id/title_dialog"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="What method do you choose?"
-        android:textColor="@color/green_dark"
-        android:textSize="20sp"
-        android:textStyle="bold"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        android:textAlignment="center"/>
-
-    <ImageButton
-        android:id="@+id/ib_camera"
-        android:layout_width="80dp"
-        android:layout_height="80dp"
-        android:layout_marginHorizontal="16dp"
-        android:layout_marginTop="24dp"
-        android:background="@drawable/ic_camera_24"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/title_dialog" />
-
-    <ImageButton
-        android:id="@+id/ib_galery"
-        android:layout_width="80dp"
-        android:layout_height="80dp"
-        android:layout_marginHorizontal="16dp"
-        android:layout_marginTop="24dp"
-        android:layout_weight="3"
-        android:background="@drawable/ic_insert_photo_24"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/title_dialog" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
