--- app/src/main/res/navigation/mobile_navigation.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/mobile_navigation"
-    app:startDestination="@+id/nav_home">
-
-    <fragment
-        android:id="@+id/nav_home"
-        android:name="com.example.capstone.ui.home.HomeFragment"
-        android:label="@string/menu_home"
-        tools:layout="@layout/fragment_home" />
-
-    <fragment
-        android:id="@+id/nav_history"
-        android:name="com.example.capstone.ui.history.HistoryFragment"
-        android:label="@string/menu_gallery"
-        tools:layout="@layout/fragment_history" />
-
-    <fragment
-        android:id="@+id/nav_favorite"
-        android:name="com.example.capstone.ui.favorite.FavoriteFragment"
-        android:label="@string/menu_slideshow"
-        tools:layout="@layout/fragment_favorite" />
-</navigation>
